Get correct hex value from rgb

+++ color/color.go
@@ -5,5 +5,5 @@ import (
 )
 
 func HexValue(c color.RGBA) uint32 {
-	return uint32(c.A<<24 | c.R<<16 | c.G<<8 | c.B)
+	return uint32(uint32(c.A<<24) | uint32(c.R)<<16 | uint32(c.G)<<8 | uint32(c.B))
 }
